Prod (#94)
* All changes for backend

* Frontend Deployed

* Backend Deployed

* Updating gitignore

* Resolved frontend bug where you are unable to create a stats log for a vehicle

* Created SSL and updated to use DNS with Route53

* Setup local development env

* Awaiting on SSL from Frontend

* Updated Changes for Localhost env

* Updated SSL and fully deployed backend

* Reworked backend for prod

* Reworked frontend for production

* Protecting endpoint testing

* Added rate limiting on vehicles endpoints

* Additional rate limiting for all API endpoints and removed API Docs access

* Fixed UI for Login and Sign up. Added toast notifications when hitting protected route without authorization

* Resolved issue with white text on newsletter subscription

* Frontend prod deployed, updated SSL

* Updated prod env code

* Resolving unnecessary API calls

* Updated Prod env

* Updated Prod yaml

* Added API for storing subscriber emails

* Added simple toast when someone signs up for newsletter - this does NOT use the API at the moment

* Updated cancel button on bug report for it to navigate back to main page when pressed, updated text for garage home to inform users where vehicle detail page is located

* Resolved issue where a user go to garage and there is not a vehicle created it would cause a 500 error

* Reworked maintenance log frontend and backend

* Resolved issue with year selector

* Updated edit feature in the backend

* Added UpdateVehicleDetail page, added the edit functionality on vehicle details page

* Added security - when a user try to access a vehicle not created by them it will redirect them to the main page with an error

* Worked on vehicle maintenance logs for the backend

* Updated the garage page and added service log page

* UI changes for maintenance log form

* Large update - added delete function for maintenance log, a dedicated page for specific maintenance log

* Added edit function for maintenance log

* Aws test

---------

+++ api/routers/vehicles.py
@@ -8,10 +8,12 @@ from fastapi import (
 )
 from typing import Union, List
 from queries.vehicles import VehicleIn, VehicleRepository, VehicleOut, Error
+from fastapi.responses import JSONResponse
 from pydantic import ValidationError
 from utils.authentication import try_get_jwt_user_data
 from models.jwt import JWTUserData
 from config import verify_api_host, oauth2_scheme
+from main import limiter
 
 
 tags_metadata = [
@@ -28,7 +30,9 @@ router = APIRouter(tags=["Vehicles"])
     response_model=Union[VehicleOut, Error],
     dependencies=[Depends(verify_api_host), Depends(oauth2_scheme)],
 )
+@limiter.limit("5/minute")
 def create_vehicle(
+    request: Request,
     vehicle: VehicleIn,
     repo: VehicleRepository = Depends(),
     current_user: JWTUserData = Depends(try_get_jwt_user_data),
@@ -60,7 +64,9 @@ def create_vehicle(
     response_model=VehicleOut,
     dependencies=[Depends(verify_api_host), Depends(oauth2_scheme)],
 )
+@limiter.limit("20/minute")
 async def get_vehicle_by_id(
+    request: Request,
     response: Response,
     vehicle_id: int,
     vehicle_repo: VehicleRepository = Depends(),
@@ -76,6 +82,11 @@ async def get_vehicle_by_id(
                 "error": f"Vehicle with ID {vehicle_id} not found",
             },
         )
+    if vehicle.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You do not have permission to access this vehicle.",
+        )
     return vehicle
 
 
@@ -84,7 +95,9 @@ async def get_vehicle_by_id(
     response_model=Union[List[VehicleOut], Error],
     dependencies=[Depends(verify_api_host), Depends(oauth2_scheme)],
 )
+@limiter.limit("20/minute")
 def list_vehicles(
+    request: Request,
     repo: VehicleRepository = Depends(),
     current_user: JWTUserData = Depends(try_get_jwt_user_data),
 ):
@@ -93,17 +106,24 @@ def list_vehicles(
 
     try:
         vehicles = repo.get_vehicles_by_user_id(current_user.id)
+        if not vehicles:  # Check if the vehicle list is empty
+            return JSONResponse(
+                status_code=200,
+                content={"message": "No vehicles in the garage"},
+            )
         return vehicles
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
 
 @router.put(
-    "/vehicles/{vehicle_id}",
+    "/vehicles/update/{vehicle_id}",
     response_model=Union[VehicleOut, Error],
     dependencies=[Depends(verify_api_host), Depends(oauth2_scheme)],
 )
+@limiter.limit("5/minute")
 async def update_vehicle(
+    request: Request,
     response: Response,
     vehicle_id: int,
     vehicle: VehicleIn,
@@ -112,13 +132,27 @@ async def update_vehicle(
 ) -> Union[VehicleOut, Error]:
     if not current_user:
         raise HTTPException(status_code=401, detail="Unauthorized")
+
+    existing_vehicle = vehicle_repo.get_vehicle_by_id(vehicle_id)
+
+    if existing_vehicle is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Vehicle with ID {vehicle_id} not found",
+        )
+
+    if existing_vehicle.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You do not have permission to update this vehicle.",
+        )
     try:
         updated_vehicle = vehicle_repo.update_vehicle(vehicle_id, vehicle)
         if updated_vehicle:
             return updated_vehicle
         else:
             raise HTTPException(
-                status_code=500, detail="Failed to update vehicle"
+                status_code=404, detail=f"Vehicle ID {vehicle_id} not found"
             )
     except ValidationError as e:
         raise HTTPException(status_code=400, detail=str(e))
@@ -127,16 +161,18 @@ async def update_vehicle(
         response.status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
         return Error(
             detail="Internal server error",
-            message=f"Vehicle {vehicle_id} does not exist.",
+            message=f"Failed to update vehicle ID {vehicle_id} due to an error.",
         )
 
 
 @router.delete(
-    "/vehicles/{vehicle_id}",
+    "/vehicles/delete/{vehicle_id}",
     response_model=Union[VehicleOut, Error],
     dependencies=[Depends(verify_api_host), Depends(oauth2_scheme)],
 )
+@limiter.limit("5/minute")
 async def delete_vehicle(
+    request: Request,
     response: Response,
     vehicle_id: int,
     vehicle_repo: VehicleRepository = Depends(),
@@ -144,6 +180,20 @@ async def delete_vehicle(
 ) -> Union[VehicleOut, Error]:
     if not current_user:
         raise HTTPException(status_code=401, detail="Unauthorized")
+    existing_vehicle = vehicle_repo.get_vehicle_by_id(vehicle_id)
+
+    if existing_vehicle is None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Vehicle with ID {vehicle_id} not found",
+        )
+
+    # Check if the current user is the owner of the vehicle
+    if existing_vehicle.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You do not have permission to delete this vehicle.",
+        )
     try:
         deleted_vehicle = vehicle_repo.delete_vehicle(vehicle_id)
         if deleted_vehicle:
@@ -163,9 +213,12 @@ async def delete_vehicle(
 
 @router.get(
     "/vehicles/user/{user_id}",
-    response_model=List[VehicleOut],
+    response_model=List[
+        VehicleOut
+    ],  # Ensure this matches your expected output
     dependencies=[Depends(verify_api_host), Depends(oauth2_scheme)],
 )
+@limiter.limit("20/minute")
 async def get_vehicles_by_user_id(
     request: Request,
     response: Response,
@@ -179,15 +232,13 @@ async def get_vehicles_by_user_id(
         vehicles = vehicle_repo.get_vehicles_by_user_id(user_id)
         return vehicles
     except HTTPException as http_exc:
-        if http_exc.status_code == 404:
-            raise
-        else:
-            print(
-                f"Failed to grab USER ID {user_id} due to an error: ",
-                http_exc.detail,
-            )
-            response.status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
-            raise HTTPException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-                detail="Internal server error",
-            )
+        print(
+            f"Failed to grab USER ID {user_id} due to an HTTP error: {http_exc.detail}"
+        )
+        raise
+    except Exception as exc:
+        print(f"Server error occurred: {exc}")
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
+            detail="Internal server error",
+        )
